requirements.txt
pandas
numpy
scipy
matplotlib
seaborn
plotly
yfinance
statsmodels
scikit-learn
xgboost
lightgbm
shap
python-dotenv
requests
beautifulsoup4
pyyaml
joblib
pyarrow
